ci: stabilize release sbom and rls grants
