+++ .github/actions/build-attested-image/action.yml
@@ -52,6 +52,8 @@ runs:
 
     - name: Generate Image SBOM
       uses: anchore/sbom-action@e22c389904149dbc22b58101806040fa8d37a610 # v0.24.0
+      env:
+        SYFT_PLATFORM: linux/amd64
       with:
         image: ${{ inputs.registry }}/${{ inputs.image-name }}@${{ steps.build.outputs.digest }}
         format: spdx-json
